requirements.txt
pandas 
spotipy
python-dotenv